Clarkeni/perp logging (#265)
* Perp instruction logging
* Onchain funding calculation
* perp_spot_transfers field on perp positions
* Logging on register token, perp market
* Additional fields on update token index logging
* maker and taker volume fields on perp position

+++ programs/mango-v4/src/instructions/perp_consume_events.rs
@@ -63,7 +63,7 @@ pub fn perp_consume_events(ctx: Context<PerpConsumeEvents>, limit: usize) -> Res
                             emit_perp_balances(
                                 ctx.accounts.group.key(),
                                 fill.maker,
-                                perp_market.perp_market_index as u64,
+                                perp_market.perp_market_index,
                                 ma.perp_position(perp_market.perp_market_index).unwrap(),
                                 &perp_market,
                             );
@@ -103,14 +103,14 @@ pub fn perp_consume_events(ctx: Context<PerpConsumeEvents>, limit: usize) -> Res
                                     emit_perp_balances(
                                         ctx.accounts.group.key(),
                                         fill.maker,
-                                        perp_market.perp_market_index as u64,
+                                        perp_market.perp_market_index,
                                         maker.perp_position(perp_market.perp_market_index).unwrap(),
                                         &perp_market,
                                     );
                                     emit_perp_balances(
                                         ctx.accounts.group.key(),
                                         fill.taker,
-                                        perp_market.perp_market_index as u64,
+                                        perp_market.perp_market_index,
                                         taker.perp_position(perp_market.perp_market_index).unwrap(),
                                         &perp_market,
                                     );
